Weather div finished with imageDisplay and date working
I did not finish the weatherDiv as I forgot to finish the date and images, which I have now done. I have come to the conclusion that my weatherApp is oriented to the spanish language, maybe I should do a button in order to change the language.

Next thing to do is now the forecast